Fixed handlebars version for the no-server example
The no-server example requires handlebars-1.0.rc.1.js, but
handlebars-1.0.0.js was included. I just replaced the latter with the
former.
